+++ app/Http/Controllers/MonsterController.php
@@ -35,16 +35,31 @@ class MonsterController extends Controller
             'monster.id' => ['numeric'],
             'monster.name' => ['required', 'string'],
             'monster.image' => ['required', 'string'],
+            'monster.hp' => ['numeric', 'required'],
             'monster.reward_xp' => ['numeric', 'required'],
             'monster.reward_gold' => ['numeric', 'required'],
     
         ]);
+
+        $data['monster']['hp'] = max($data['monster']['hp'], 0); 
+        $data['monster']['hp'] = min($data['monster']['hp'], 100); 
 
         $monster = Monster::find($data['monster']['id']);
         if($monster->classroom_id != $class->id)
             abort(403, 'What are you trying? :(');
         
         $monster->update($data['monster']);
+    }
+
+    public function battle() {
+        $monster = Monster::findOrFail(request()->id);
+        $class = Classroom::findOrFail($monster->classroom_id);
+        $this->authorize('update', $class);
+
+        $monster->update([
+            'hp' => max(0 , $monster->hp + request()->value), 
+        ]);
+
     }
 
     public function store($code) {