+++ resources/lang/ca/monsters.php
@@ -14,5 +14,6 @@ return [
       'reward' => 'Recompenses per a cada estudiant al vèncer',
       'reward_xx' => 'Recompensa ',
       'image' => 'Imatge',
+      'hp' => 'Punts de vida restants',
 
 ];