+++ resources/lang/en/monsters.php
@@ -14,5 +14,6 @@ return [
       'reward' => 'Reward for every student when winning',
       'reward_xx' => 'Reward ',
       'image' => 'Image',
+      'hp' => 'Remaining health points',
 
 ];