+++ resources/views/studentsview/index.blade.php
@@ -1,7 +1,7 @@
 @extends('layouts.student')
 
 @section('content')
-       <show-students-student :character-theme="{{ $class->character_theme }}" :students="{{ json_encode($students) }}" :theme="{{ $class->theme }}"></show-students-student>
+       <show-students-student :rating="{{ $rating }}" :monsters="{{ $monsters }}" :character-theme="{{ $class->character_theme }}" :students="{{ json_encode($students) }}" :theme="{{ $class->theme }}"></show-students-student>
 @endsection
 
 @push('scriptchat')